need to finish dl files

+++ app/Http/Controllers/ProcessesController.php
@@ -27,19 +27,25 @@ class ProcessesController extends Controller
 
     }
     public function addFiles(Request $request,$n,$dir,$process){
-        $file = $request->file('file'.$n);
-        $path = $file->store($dir);
-        $array = explode('/', $path,2);
-        $newfilename = "/file".$n.".pdf";
-        $newfilepath = $dir . $newfilename;
-        if(Storage::disk('local')->exists($dir.$newfilename)){
-            Storage::delete($newfilepath);
+        $type = "/pdf/";
 
+        $file = $request->file('file'.$n);
+        if(preg_match($type,$file->getMimeType())) {
+            $path = $file->store($dir);
+            $array = explode('/', $path, 2);
+            $newfilename = "/file" . $n . ".pdf";
+            $newfilepath = $dir . $newfilename;
+            if (Storage::disk('local')->exists($dir . $newfilename)) {
+                Storage::delete($newfilepath);
+
+            }
+            $filename = substr($array[1], strlen('/files/process_' . $process->id));
+            Storage::move($dir . "/" . $filename, $newfilepath);
+            //return $newfilepath;
+            return $newfilepath;
+        } else {
+            return $newfilepath = False;
         }
-        $filename = substr($array[1], strlen('/files/process_' . $process->id));
-        Storage::move($dir . "/" . $filename, $newfilepath);
-        //return $newfilepath;
-        return $newfilepath;
     }
 
 
@@ -47,6 +53,7 @@ class ProcessesController extends Controller
     public function uploadFiles(Request $request)
     {
         $process = Process::find(1);
+
         if ($request->file1 or $request->file2 or $request->file3 or $request->file4 or $request->file5 or $request->file6) {
             $dir='public/files';
             if (Storage::disk('local')->exists($dir)){
@@ -59,43 +66,66 @@ class ProcessesController extends Controller
             } else {
                 Storage::makeDirectory($dir);
             }
-            $files = $process->file;
-            $filesarr = explode('"-"', $files);
+            $files = $process->file; #VERIFICAR
+            $filesindb = explode('"-"', $files);
+
+            }
             if ($request->file1) {
-                $file = $this->addFiles($request, '1', $dir, $process);
-                $filesarr[0]=$file;
+                $file[0] = $this->addFiles($request, '1', $dir, $process);
             }
             if ($request->file2) {
-                $file = $this->addFiles($request, '2', $dir, $process);
-                $filearr[1] = $file;
+                $file[1] = $this->addFiles($request, '2', $dir, $process);
             }
             if ($request->file3) {
-                $file = $this->addFiles($request, '3', $dir, $process);
-                $filearr[2] = $file;
+                $file[2] = $this->addFiles($request, '3', $dir, $process);
             }
             if ($request->file4) {
-                $file = $this->addFiles($request, '4', $dir, $process);
-                $filearr[3] = $file;
+                $file[3] = $this->addFiles($request, '4', $dir, $process);
             }
             if ($request->file5) {
-                $file = $this->addFiles($request, '5', $dir, $process);
-                $filearr[4] = $file;
+                $file[4] = $this->addFiles($request, '5', $dir, $process);
             }
             if ($request->file6) {
-                $file = $this->addFiles($request, '6', $dir, $process);
-                $filearr[5] = $file;
+                $file[5] = $this->addFiles($request, '6', $dir, $process);
             }
-            $fileaux = "";
-            foreach ($filesarr as $file)
-                if ($fileaux == "") {
-                    $fileaux = $file;
-                } else {
-                    $fileaux = $fileaux . '"-"' . $file;
-                }
-            $process->file = $fileaux;
-            $process->save();
 
-        }
+            $fileaux = array();
+
+            foreach ($file as $f) {
+                if($f != False) {
+                    if (in_array($f, $fileaux)) {
+                    } else {
+                        array_push($fileaux, $f);
+                    }
+                } else {
+                    return redirect ('/uploadfiles')->with('notPDF',True);
+                }
+            }
+
+            foreach ($filesindb as $fa) {
+                if (in_array($fa, $fileaux)) {
+                } else {
+                    array_push($fileaux, $fa);
+                }
+            }
+            $files = "";
+            foreach ($fileaux as $file) {
+                if ($files == "") {
+                    $files = $file;
+                } else {
+                    if($file != null) {
+                        $files = $file . '"-"' . $files;
+                    } else {
+                        $files = $files;
+                    }
+                }
+
+            }
+            $process->file = $files;
+            $process->save();
+            return $files;
+
+
     }
     public function ChangeProperty(Request $request){
         $property = $request->change;
@@ -105,5 +135,31 @@ class ProcessesController extends Controller
         }else{
             $process->active = $request->active;
         }
+    }
+
+    public function downloadFile(Request $request){
+        $process = Process::find(1);
+        $files = $process->file;
+        $filesarr = explode('"-"', $files);
+        //'storage/files/process_1/file6.pdf'
+
+        $var = '/public\//';
+        $files = array();
+        foreach ($filesarr as $file) {
+            if(preg_match($var,$file)){
+                array_push($files,preg_replace($var,'storage/',$file));
+            }
+        }
+        $var = '/storage\/files\/process_1\/file/';
+        $values = array();
+        foreach ($files as $fil){
+            if (preg_match($var, $fil)){
+                $value = preg_replace('/.pdf/','',preg_replace($var,'',$fil));
+                array_push($values,$value);
+            }
+        }
+        return dd($values);
+
+        return view('downloadfiles',compact('files'));
     }
 }
